move impl. to cluster_config_validate
